+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
 - Initial version.
+- Multisignature coordinator for creating shared multisignature Hierarchical 
+  Deterministic (HD) wallets.